update search page
