+++ environment/template.env
@@ -3,6 +3,10 @@ POSTGRES_DB=pephub
 POSTGRES_USER=docker
 POSTGRES_PASSWORD=password
 
+QDRANT_HOST="localhost"
+QDRANT_PORT=6333
+QDRANT_ENABLED=true
+
 GH_CLIENT_ID=client_id
 GH_CLIENT_SECRET=client_secret
 REDIRECT_URI="http://localhost:8000/auth/callback"
