+++ pephub/routers/api/v1/search.py
@@ -31,7 +31,7 @@ async def search_for_pep(
     qdrant: QdrantClient = Depends(get_qdrant),
     model: SentenceTransformer = Depends(get_sentence_transformer),
     db: Connection = Depends(get_db),
-    limit: int = 20,
+    limit: int = 8,
 ):
     """
     Perform a search for PEPs. This can be done using qdrant (semantic search),
@@ -47,18 +47,31 @@ async def search_for_pep(
                 query_vector=query_vec,
                 limit=limit,
             )
-            return JSONResponse(
-                content={
-                    "query": query.query,
-                    "results": [r.dict() for r in results],
-                    "namespace_hits": list(
+            namespaces = db.get_namespaces_info_by_list()
+            namespace_hits = [
+                n.namespace
+                for n in namespaces
+                if query.query.lower() in n.namespace.lower()
+            ]
+            namespace_hits.extend(
+                [
+                    n
+                    for n in list(
                         set(
                             [
                                 r.dict()["payload"]["registry"].split("/")[0]
                                 for r in results
                             ]
                         )
-                    ),
+                    )
+                    if n not in namespace_hits
+                ]
+            )
+            return JSONResponse(
+                content={
+                    "query": query.query,
+                    "results": [r.dict() for r in results],
+                    "namespace_hits": namespace_hits,
                 }
             )
         except Exception as e:
@@ -70,7 +83,7 @@ async def search_for_pep(
         for namespace in namespaces:
             try:
                 res = db.search.project(
-                    namespace=namespace, search_str=query.query, limit=limit
+                    namespace=namespace.namespace, search_str=query.query, limit=limit
                 )
                 results.extend(res.results)
             except:
