+++ pephub/static/templates/namespace_search_error.html
