+++ pephub/static/templates/search_spinner.html
@@ -1,4 +1,4 @@
-<div class="d-flex flex-row align-items-center justify-content-center mt-2">
+<div class="d-flex flex-row align-items-center justify-content-center mt-5">
     <div class="spinner-border text-secondary" role="status">
       <span class="visually-hidden">Loading...</span>
     </div>
